backend/requirements.txt
flask
flask-cors
google-generativeai
praw
bertopic
